feat(recipes): Implement recipe management with backend integration
- Prevent editing of collected recipes
- Add route protection for recipe editing
- Fix dropdown menu styling and positioning
- Improve user experience with proper navigation
- Prepare for backend integration by removing frontend ID management

BREAKING CHANGES TODO:
- Recipe IDs will be managed by backend instead of frontend

+++ front-end/src/app/recipes/recipe-detail/recipe-detail.component.ts
@@ -12,27 +12,27 @@ import { RecipeService } from '../recipe.service';
 export class RecipeDetailComponent implements OnInit {
   recipe: Recipe;
   id: number;
-  isMyRecipe: boolean;
-
-  constructor(private recipeService: RecipeService,
-              private route: ActivatedRoute,
-              private router: Router) {
-  }
+  isInMyRecipes: boolean = false;
+  isPersonal: boolean;
+  constructor(
+    private recipeService: RecipeService,
+    private route: ActivatedRoute,
+    private router: Router
+  ) {}
 
   ngOnInit() {
-    // Check if we're in the my-recipes section
-    this.isMyRecipe = this.router.url.includes('/my-recipes');
-    
-    this.route.params
-      .subscribe(
-        (params: Params) => {
-          this.id = +params['id'];
-          this.recipe = this.recipeService.getRecipe(this.id);
-          if (!this.recipe) {
-            this.router.navigate(['/recipe-book']);
-          }
+    this.route.params.subscribe(
+      (params: Params) => {
+        this.id = +params['id'];
+        this.recipe = this.recipeService.getRecipe(this.id);
+        this.isPersonal =  this.recipe.isPersonal;
+        if (!this.recipe) {
+          this.router.navigate(['/recipe-book']);
+        } else {
+          this.isInMyRecipes = this.recipeService.isInMyRecipes(this.recipe);
         }
-      );
+      }
+    );
   }
 
   onAddToShoppingList() {
@@ -40,22 +40,26 @@ export class RecipeDetailComponent implements OnInit {
   }
 
   onEditRecipe() {
-    this.router.navigate(['edit'], {relativeTo: this.route});
-    // this.router.navigate(['../', this.id, 'edit'], {relativeTo: this.route});
+    if (this.recipe.isPersonal) {
+      this.router.navigate(['edit'], {relativeTo: this.route});
+    }
   }
 
   onDeleteRecipe() {
-    this.recipeService.deleteRecipe(this.id);
-    this.router.navigate(['/recipes']);
+    if (this.recipe.isPersonal) {
+      this.recipeService.deleteRecipe(this.id);
+      this.router.navigate(['/recipes']);
+    }
   }
 
-  onSubmitAsIdea() {
-    // Create a copy of the current recipe
-    const currentRecipe = { ...this.recipe };
-    
-    // Add to recipe ideas and navigate
-    this.recipeService.addRecipe(currentRecipe);
-    this.router.navigate(['/recipe-ideas']);
+  onToggleFavorite() {
+    if (!this.recipe.isPersonal) {
+      if (this.isInMyRecipes) {
+        this.recipeService.removeFromMyRecipes(this.recipe);
+      } else {
+        this.recipeService.addToMyRecipes(this.recipe);
+      }
+      this.isInMyRecipes = !this.isInMyRecipes;
+    }
   }
-
 }
